requirements.txt
certifi==2018.8.24
chardet==3.0.4
falcon==1.4.1
hug==2.4.1
idna==2.7
numpy==1.15.1
oauthlib==2.1.0
PySocks==1.6.8
python-mimeparse==1.6.0
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.23

git+https://github.com/tweepy/tweepy
